+++ src/components/BotDetails/BotDetailsNavBar.tsx
@@ -3,7 +3,7 @@ import {View, Image, TouchableOpacity, Clipboard, StyleSheet} from 'react-native
 import {observer, inject} from 'mobx-react/native'
 import {k} from '../Global'
 import {colors} from '../../constants'
-import {IBot, IWocky} from 'wocky-client'
+import {IBot} from 'wocky-client'
 import {RText} from '../common'
 // import AddBotPost from './AddBotPost'
 import {Actions} from 'react-native-router-flux'
@@ -11,29 +11,18 @@ import {isAlive} from 'mobx-state-tree'
 import {navBarStyle} from '../Router'
 
 type Props = {
-  botId: string
-  server?: string
-  isNew: boolean
-  params?: any
-  wocky?: IWocky
-  analytics?: any
-  scrollable: boolean
+  bot: IBot
+  notificationStore?: any
 }
 
-const LocationDetailsNavBar = inject('wocky')(({wocky, botId, server}: Props) => {
-  const bot: IBot = wocky.getBot({id: botId, server})
-  return <Header bot={bot} />
-})
-
-const Header = inject('notificationStore')(
-  observer(({bot, scale, notificationStore}) => {
-    const map = scale === 0
+const BotDetailsNavBar = inject('notificationStore')(
+  observer(({bot, notificationStore}: Props) => {
     if (!bot || !isAlive(bot)) {
       return null
     }
     const {backButtonImage, navBarButtonColor, titleStyle} = navBarStyle
     return (
-      <View style={[styles.title]}>
+      <View style={styles.title}>
         <TouchableOpacity onPress={Actions.pop}>
           <Image
             source={backButtonImage}
@@ -51,7 +40,7 @@ const Header = inject('notificationStore')(
           style={{marginHorizontal: 16}}
         >
           <RText
-            numberOfLines={map ? 1 : 2}
+            numberOfLines={2}
             // must wait for solution to https://github.com/facebook/react-native/issues/14981
             // adjustsFontSizeToFit
             minimumFontScale={0.8}
@@ -66,18 +55,16 @@ const Header = inject('notificationStore')(
           >
             {bot.error ? 'Bot Unavailable' : bot.title}
           </RText>
-          {map && (
-            <RText
-              minimumFontScale={0.6}
-              numberOfLines={1}
-              weight="Light"
-              size={14}
-              color={colors.DARK_PURPLE}
-              style={{textAlign: 'center'}}
-            >
-              {bot.address}
-            </RText>
-          )}
+          <RText
+            minimumFontScale={0.6}
+            numberOfLines={1}
+            weight="Light"
+            size={14}
+            color={colors.DARK_PURPLE}
+            style={{textAlign: 'center'}}
+          >
+            {bot.address}
+          </RText>
         </TouchableOpacity>
         <View style={{marginRight: 20 * k}}>
           <ShareButton bot={bot} />
@@ -100,17 +87,18 @@ const ShareButton = observer(({bot}) => {
   ) : null
 })
 
-export default LocationDetailsNavBar
+export default BotDetailsNavBar
 
 const styles = StyleSheet.create({
   title: {
     height: 64,
     flex: 1,
+    backgroundColor: 'white',
     elevation: 1,
     paddingTop: 20,
     alignItems: 'center',
     justifyContent: 'space-between',
     flexDirection: 'row',
-    marginHorizontal: 10,
+    paddingHorizontal: 10,
   },
 })