+++ src/components/Router.tsx
@@ -24,10 +24,8 @@ import MyAccount from './MyAccount'
 import ProfileDetail from './ProfileDetail/ProfileDetail'
 import ChatListScreen from './ChatListScreen'
 import ChatScreen from './ChatScreen'
-// import BotCompose from './BotCompose'
-// import BotCreate from './map/BotCreate'
-import LocationDetails from './LocationDetails'
-import LocationDetailsNavBar from './LocationDetails/LocationDetailsNavBar'
+import BotDetails from './BotDetails/BotDetails'
+import BotDetailsNavBar from './BotDetails/BotDetailsNavBar'
 import BotsScreen from './BotsScreen'
 // import ExploreNearBy from './map/ExploreNearBy'
 import TestRegister from './TestRegister'
@@ -100,24 +98,12 @@ export const navBarStyle = {
   },
 }
 
-// const tinyRobotTitle = () => (
-//   <TouchableOpacity onPress={() => Actions.refs.home.scrollToTop()}>
-//     <Text style={dayNavBar.titleStyle}>tinyrobot</Text>
-//   </TouchableOpacity>
-// )
-
 const iconClose = require('../../images/iconClose.png')
 // const baseMessagesIcon = require('../../images/iconMessage.png')
 // const newMessagesIcon = require('../../images/newMessages.png')
 const sendActive = require('../../images/sendActive.png')
 
 const uriPrefix = settings.isStaging ? 'tinyrobotStaging://' : 'tinyrobot://'
-
-// TODO: is it still necessary for react-navigation 2.x ?
-// prevent keyboard from persisting across scene transitions
-// autorun(() => {
-//   if (Actions.currentScene !== '') Keyboard.dismiss()
-// })
 
 type Props = {
   wocky?: IWocky
@@ -143,12 +129,12 @@ class TinyRobotRouter extends React.Component<Props> {
       {delay: 1000}
     )
 
-    // TODO: run locationPrimer from fullMap ? Actions.currentScene is not reactive anymore
-    // autorun(() => {
-    //   if (Actions.currentScene === '_fullMap' && !locationStore!.alwaysOn && !store.locationPrimed) {
-    //     if (Actions.locationPrimer) Actions.locationPrimer()
-    //   }
-    // }, {delay: 1000} )
+    autorun(() => {
+      const {navStore: {scene}, store: {locationPrimed}, locationStore: {alwaysOn}} = this.props
+      if (scene === 'home' && !alwaysOn && !locationPrimed) {
+        if (Actions.locationPrimer) Actions.locationPrimer()
+      }
+    }, {delay: 1000} )
   }
 
   render() {
@@ -216,7 +202,7 @@ class TinyRobotRouter extends React.Component<Props> {
               </Stack>
               <Scene key="bottomMenu" component={BottomMenu} />
               <Scene key="createBot" path="bot/:server/:botId/:params*" component={CreationHeader} fromTop />
-              <Scene key="botDetails" path="bot/:server/:botId/:params*" component={LocationDetails} draggable opacityHeader={LocationDetailsNavBar} />
+              <Scene key="botDetails" path="bot/:server/:botId/:params*" component={BotDetails} draggable opacityHeader={BotDetailsNavBar} />
               <Scene key="botCompose" component={BotCompose} back />
               <Scene key="botEdit" component={BotCompose} edit back />
               <Scene key="camera" component={Camera} />